requirements.txt
autopep8==1.4.4
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5  # django
django-cors-headers==3.2.1  # django cors headers
django-rest-passwordreset==1.1.0  # reset senha
djangorestframework==3.11.0  # django rest framework

entrypoints==0.3  # flake8
flake8==3.7.9  # flake8
flake8-django==0.0.4  # flake8
pycodestyle==2.5.0  # flake8
mccabe==0.6.1  # flake8
pyflakes==2.1.1  # flake8

Pillow==7.1.1  # foto do usuario
reportlab==3.5.42
webencodings==0.5.1
html5lib==1.0.1
PyPDF2==1.26.0
six==1.15.0
xhtml2pdf==0.2.4

requests==2.24.0
sqlparse==0.3.1
whitenoise==5.0.1

pep8==1.7.1
pytz==2019.3
sqlparse==0.3.1
gunicorn==20.0.4

psycopg2==2.8.5